% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.super.R
\name{addSupClust}
\alias{addSupClust}
\title{Add a supercluster layer}
\usage{
addSupClust(map, path = "")
}
\arguments{
\item{map}{a map widget object created from \code{leaflet::leaflet()}}

\item{path}{relative path / absolute URL to an Arrow IPC stream file}
}
\description{
Add a supercluster layer
}
\examples{
\dontrun{
map <- leaflet() \%>\%
  addProviderTiles(provider = providers$CartoDB.DarkMatter) \%>\%
  addSupClust(
    path = "data/big.arrow", ~lon, ~lat, layerId = ~ID, radius = 4, stroke = FALSE,
    fillOpacity = 0.8, label = ~paste("ID: ", ID, sep = ""),
    popupOptions = popupOptions(closeButton = FALSE),
    clusterOptions = markerClusterOptions()
  ) \%>\%
  setView(lng = 0, lat = 0, zoom = 2)
}
}
